add utils and .whl file
