+++ setup.py
@@ -1,0 +1,8 @@
+from setuptools import setup, find_packages
+
+setup(
+    name="fabric_utils",
+    version="0.1.0",
+    packages=find_packages(),
+    install_requires=[],  # spark deps are already there in Fabric
+)
